Rename of Project
Renamed project from "CPR - Timers" to "CPR-Scouting-App"
