+++ .idea/.name
@@ -1,1 +1,1 @@
-CPR - Timers
+CPR-Scouting-App
