+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">CPR - Timers</string>
+    <string name="app_name">CPR-Scouting-App</string>
     <string name="button_start_match">Start Match</string>
     <string name="button_start_teleop">Start Teleop</string>
     <string name="button_end_match">End Match</string>
