+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.CPRTimers" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.CPRScoutingApp" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.CPRTimers" parent="Base.Theme.CPRTimers" />
+    <style name="Theme.CPRScoutingApp" parent="Base.Theme.CPRScoutingApp" />
 </resources>
